+++ changelog.md
@@ -18,10 +18,16 @@
 * Added basic support for motion broadcastin
 
 ### 2017-06-28
+[Lidefeath]
+* Added linked generators i.e. a generator can spawn other generators which then spawn objects - theoretically this can go even deeper
+* Added testdata for the linked generators near Holtburg: 
+	- spawn a Drudge Skulker with his generator
+	- spawn a Drudge Camp (3 drudges) with its generator
+	- randomly spawn 10 drudge camp generators in a landblock
+
 [Jyrus]
 * Implement Auto close timer for Doors
 
-### 2017-06-28
 [Og II]
 * Update readme with db minimum required versions.